requirements.txt
PyMySQL
pycryptodome
mysql.connector
python-decouple
Flask
tk